src/main/java/com/msampietro/springspecquerylanguage/predicatebuilder/LessThanOrEqualPredicate.java
package com.msampietro.springspecquerylanguage.predicatebuilder;

import com.fasterxml.jackson.databind.ObjectMapper;

import javax.persistence.criteria.CriteriaBuilder;
import javax.persistence.criteria.CriteriaQuery;
import javax.persistence.criteria.Predicate;
import javax.persistence.criteria.Root;
import java.io.Serializable;

public class LessThanOrEqualPredicate<T> extends BasePredicate<T> {

    LessThanOrEqualPredicate(ObjectMapper objectMapper, Class<? extends Serializable> idClazz) {
        super(objectMapper, idClazz);
    }

    @SuppressWarnings("NullableProblems")
    @Override
    public Predicate toPredicate(Root<T> root, CriteriaQuery<?> query, CriteriaBuilder builder) {
        return builder.lessThanOrEqualTo(getCriteriaExpressionKey(root), parseValue(getCriteriaObjectValue().toString(), builder));
    }
}
